Merged sphere.h into shape.h
All shapes and future shape definitions, typedefs and prototypes to be added to shape.h directly.

The primary cause for this was for the union in the s_shape structure that needs to know the full size of each element before hand.

sphere_intersect.c now includes shapes.h instead of sphere.h

Added forward declaration section to shape.h to include:
    - t_intersect_report
    - t_ray