New function: add_new_shape_to_world
int	add_new_shape_to_world(t_world *world, t_shape_type type);

Creates a new instance of a default shape of the type provided.

Creates a new node containing this shape (content) and adds it to the back of the shapes list present in the t_world structre.

Returns a non-zero value if an error occurs. Returns 0 on success.