add_intersects_to_list:	bug fix
Added interation over report.count.

Previously no iteration (and correction of the report.t indexation) was
being made. Miraculously this was avoiding crashes and working
somehow...

This would cause only one intersection to ever be added (when two occur
for a sphere) to the list, causing strange behaviour in regards to the
lighting.