New intersection util: get_first_hit()
In new file: src/intersection/utils.c

Acts as a wrapper for dbl_lstgetone that passes is_hit_positive() as an
argument.

Reason: Clarity of use when it will be used in 'color_at()' [not yet
defined], when we want to get the 'first_hit' in a given list of
intersections.

first_hit = get_first_hit(&intersects);

is more readdable than

first_hit = dlb_lstgetone(&intersects, is_hit_positive);