New file: intersect_world.c
Changes to headers:

	- minirt.h: new define: INVALID_POINTER
	- intersection.h: new define: INTERSECT_INCOMPLETE

intersect_world.c:

___

core function: intersect_world()

Returns a list of all the intersections found in the world with the provided ray.

____

sub-functions:

	-	intersect_all_shapes()
populates the list with all intersections found between the given ray & list of shapes.

	-	add_intersects_to_list()
populates the list with any intersections detected (in the report)

	-	reset_report()
performs a bzero (ft_bzero) on the report structure (readability).