New function: generate_image (rendering)
render now calls generate_image to generate an image with raycasting.

generate_image iterates over the entire display (window) and calculates
the color for each pixel at stores that data in the corresponding memory
slot of the mlx_image.

If color_at returns black (0 for all values) no call is made to
set_pixel_color as the values in the image are all set to zero by
default.

Currently returns 0 in all cases.

Ready to return a non-zero value if/when error-handling is implemented in
the color_at() function.