Parser: added info validation utilities
First prototypes of the following:

- validate_brightness
- validate_color
- validate_coordinate
- validate_vector
- validate_fov
- validate_size

In theory they all are capable of validating a string and logging any
errors encountered.

These have been tested in an external environment (without error logging
capability) and were functionnal. Testing is required within the miniRT
program.

The calls to these are to be made iteratively according to expected
order from:

- validate_ambient
- validate_camera
- validate_light
- validate_sphere
- validate_plane
- validate_cylinder

^ (TODO) ^