libft: fixed a bug in ft_atod()
The return value of increment() was incorrectly being assigned to
number.

The sub-function is now being called correctly and the value of number
is updated from within increment().

Changed name of function from ft_atof => ft_atod to reflect the use of
doubles (to avoid loss of accuracy).

Changes reflected elsewhere in miniRT program.