parsing: bug fixes
destroy_input():

	Now calls ft_lstclear() for the errors linked list. This was
previously not being done and causing leaks when parsing errors were
logged/reported.

validate_token():

	Now returns if validate_id returns any non-zero value (was
previously only -1). This was causing and invalid ID to continue to call
the function in the validate array at index -1 (value of ID_INVALID) and
causing a segmentation fault.

validate_ambient():

	Now keeps track of if an ambient element already exists and
returns an error if it does. This duplication prevention will be
mirrored for other single elements.