Camera: initialization from input data
init_camera now takes an array of strings as an argument (the camera
line found & validated in the .rt file).

The fov and view transform is now applied with this data.

A new utility function has been added to camera.c: str_to_tuple.

str_to_tuple: converts a validated string to a tuple. This implies that
the string consists of 3 valid values (x, y, z) seperated by a ','.

init_data(): init_mlx moved to end of function so that a window is only
opened once all input data has been validated and initialized.