Fixed/Renamed get_rot() => rotate_y_to()
Moved from plane.c to normal.c

Fixed a small typo in return value of matrix that was causing abnormal behaviour.

vector 'y' (previous named 'a') is now constant within the function.

vector 'b' renamed => 'to'

Reflection matrix (if dot == -1) is now around the X axis which is more comprehensible (but achieves the same result).

REMINDER:

This function will return the rotation matrix to be applied to vector 'to' so that it has the same orientation as constant vector 'y' = {0, 1, 0}.