Added new function: set_cylinder()
Same logic applied as for set_sphere & set_plane.

The transformation matrix to apply both diameter and height parameters can be applied in a single call as the diameter only affects X & Z scaling, and diameter affects the Y scaling.

The following order must be respected in regards to application of the transformation matrices:

- Scaling
- Rotation
- Translation