Added 2 new scenes
lots_of_spheres:

Lots of spheres with the light positioned in the center at 0,0,0

acne_lots_of_spheres:

Same as above but the yellow sphere (line 7 of .rt file) has its
diameter changed to 2. This means the edge of the sphere is directly on
the origin of the light source at 0,0,0. This is likely the cause of the
acne.