Fixed camera orientation bug (+ simplification)
Fixed a bug that was causing:

- 1: The camera view_transform matrix to be
distorted when the view direction was downwards.

- 2: The camera view_transform matrix to be wrong if the direction vector
  was {0, -1, 0} (directly downwards) and subsequently causing an error
  when calculating the inverse (0 determinant).

FIX 1: The problem was the 'left' vector in mx_orientation not being normalized
beore calculating the 'true_up'.

FIX 2: An exception is to be made when the direction vector is directly
downwards, defining left as {-1, 0, 0}.

Additional simplification:

view_transform & mx_orientation now take a 'forward' vector as the second
argument rather that a point 'to'. This is much more streamlined based
on the expected input from the .rt file is a vector. The previous
implementation was converting this vector to a 'to' point to then
immediately convert it back to a forward vector in mx_orientation.