Rounding of floats applied to input data
New function: rt_roundf() in init_utils.c

	Rounds a given float to the nearest value of EPSILON.
	- A value smaller than approximalely half of EPSILON will be
rounded down to 0.
	- A value above approximately (1.5 * EPSILON) would be rounded
	  up to (2 * EPSILON).

This rounding is applied to all input values that are accepted as
floating point numbers:
	- Ambient lighting ratio
	- Coordinate values
	- FOV
	- Light brightness ratio
	- Sphere diameter
	- Cylinder diameter
	- Cylinder height
	- Direction vector values

This rounding is made after the parsing and during the world
initialization.

Applying the rounding to the parsing could cause some strange cases such
as a very small negative number incorrectly being rounded to 0.