Small fixes : display
==

Removed an error message that was displaying when no "M" element was provided.

==

Added confirmation messages in get_available_core_count when attempting to detect the number of online cores for multi-threading.

It will either confirm the number of online logical cores detected (used as the number of threads to create), or display an error (root cause from sysconf) and a message stating that the default thread count has been used (as defined in DEFAULT_THREAD_COUNT).

==

Removed render thread count message in render()