Bug fix : Dealing with odd resolutions
Fixed a bug where the last few lines of the render were ommited if the
resolution height was not divisible by the thread count.

For example if using 12 cores and a resolution height of 1211, the
bottom 11 lines were not being rendered.

Quick fix: the last thread will now cover any leftover lines.

Long term fix will be to implement a tiling system for rendering and
potentially stricter resolution input parameters.