Implement intersect_cube function
New file added:
	bonus/src/shape/cube_intersect_utils.c

New functions:
	static void	check_axis(t_report *min_max, float origin, float direction);
	static float	get_largest_min(float x, float y, float z);
	static float	get_smallest_max(float x, float y, float z);
	float	ft_get_min_or_max_float(float a);

Cube intersect function will compute the t value
for each of the 6 plane of the cube and ordered as t_min and t_max for each axes.
The law to find the valid intersection is to get the largest min value
and the smallest maximum value
And if the t_min is well smaller than the t_max, the ray hit the cube.