Refactor of multi-thread error reporting
print_error now shares a mutex with update_progress().

This is handled with the print_mutex() function that can initialize,
destroy, lock and unlock the mutex.

Some calls to perror have been swapped for print_error so that all
threads utilise the mutex to print out a clear error log in case of
errors during rendering.