Debug fresnel effect computation.
Computation should be done only on the first intersection of the scene
and not recursively with the reflection and refraction depth.

New member to t_render_info structure:
	int			depth;

Before each time the function color_at is called again
in refraction_color and reflection_color the depth is incremented by one.
As we give a copy of info value, we can use it to affect
the schlick approximation computation only when the depth is 0