Multi thread error log redirect : refacto
Previous version had two issues:

- If an error occured during the redirect process, this was causing the
  stderr stream to be closed and rendering further error messages
undisplayable.

- The stderr could not be restored after the rendering (will be needed
  for later implementations).

The new version now stores 2 new FDs in the t_data structure (to be
improved with a struct) and applies the redirection(s) through
open|dup|dup2 calls.

Any errors encountered during this process are displayed with perror
rather than print_error (which would display a timestamp).

If any errors are encountered, the program exits as it can not ensure a
clean output as there would be a conflict between stdout and stderr with
the render progress update (%).

CLEANUP REQUIRED.