Camera refactoring : preparing for movement controls
Added following elements to t_camera structure:

	- t_point from
	- t_vector forward
	- t_vector up

Removed following elements from t_camera structure:

	- hsize
	- vsize

calibrate_camera() now takes the data->resolution as an argument as this is the only time the camera requires it.

With the new elements, these are now ready to be modified through the keyhooks/controls and the transformation/inverse matrices can be recalculated.