Camera control added : PITCH : UP (UP_KEY) | DOWN (DOWN_KEY)
Camera is now initialized with the help of get_true_up().

This will set the value of up based on the given forward vector (from
parsing) and using the default up (set to (0, 1, 0)).

The only exception is if the initial forward vector is directly down (0,
-1, 0), then the left is manually set to (-1, 0, 0).

When further changes are made to the camera (with view_transform after
changing the up and/or forward vectors), the get_true_up() function is
no longer required, avoiding the bug where the 'left' would occasionally
flip upon doing a full PITCH rotation.

mx_orientation() (in view_transform.c) has now been simplified as a
result.

Two new prototypes added to render.h: get_left & get_true_up.