Editor : Applied shape movement controls
Once a shape is selected, it can now be moved with the QWEASD keys in the same manner as the camera controls.

The movement will be applied to the camera forward vector:

	FORWARD (W) | BACK (S)
		Will move the shape along the camera forward axis.
	RIGHT (D) | LEFT (A)
		Will move the shape along the relative right|left axis (based on camera forward & up).
	UP (E) | DOWN (Q)
		Will move the shape along the vertical axis (camera up).

move_camera_origin() has had a lot of its functionality delegated to get_relative_movement() which now returns the movement vector to apply. This can be called for both the movement of the camera and the shapes as they are both relative to the camera settings.

Each shape type has a new function in its vtable: move

This simply applies the movement vector as a translation to the existing transformation matrix and re-calculates the inverse.