Editor : MODE_LIGHT implemented
In editor mode (E), (TAB) will now toggle between the 3 following modes:

    - CAMERA (default)
    - SHAPE_SELECT
    - LIGHT

In LIGHT mode, the controls are the same as for moving the camera or selected shape (QWEASD)

Changed the logic of switch_editor_mode to use modulo and a different maccro. The previous implementation with XOR is not compatible with 3 modes.

Bug fix :
    Any selected shape will now be correctly de-selected when toggling out of the SHAPE_SELECT mode.