Solve bug with ligting inside a shape.
This bug was present to compute shadow or reflection if the camera was
inside of the shape.

In function set_first_hit_valid(t_shape *self, t_inter_lst **first_hit) :
	Firstly we set the first hit to the next one if there was a self intersection.
	But in case the camera is inside the shape and the light is outside the shape:
		The shape should be self intersect to be in shadow.

	Finally we give a third argument to the function : bool inside
		We ignore self intersection if th ray is coming from the inside of the shape

Same modification for reflection. When calling again the color_at function,
instead of just give self argument, we give the details of the previous intersection
that contains self and the information if the previous ray was coming from inside of the shape.