Hotfix - solve edge case with cylinder intersection
in some case the same intersection could be counted for the cylinder caps and body.
That result to have more than 2 intersections for the cylinder.

In this case, in our algorithm, it try to write a float at position 2 of array t_report.t[2]
That result to a bad acces of memory

We add a check if intersection count is lower than 2, before to compute the second cap of intersect_caps() function